support volume up and volume down utterances
This should better be done by passing the current volume as sensor,
then adding/subtracting some value, and sending an absolute volume
value back to the client. This way, we would get consistent
behaviour across devices.

But for now, let us do it in this way until sensors are supported
at least in susi_linux.